Adjusting the project structure to stay sane when growing.
