+++ Examples/RabbitExpress.Example.Publisher/appsettings.json
