+++ Examples/RabbitExpress.Example.Worker/appsettings.json
