--- RabbitExpress.ExampleWorker/RabbitExpress.ExampleWorker.csproj
@@ -1,26 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-	<PackageReference Include="Microsoft.Extensions.Configuration" Version="2.2.0" />
-	<PackageReference Include="Microsoft.Extensions.Configuration.EnvironmentVariables" Version="2.2.4" />
-	<PackageReference Include="Microsoft.Extensions.Configuration.FileExtensions" Version="2.2.0" />
-	<PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="2.2.0" />
-	<PackageReference Include="RabbitExpress.JsonSerializer" Version="1.*" />
-  </ItemGroup>
-	
-  <ItemGroup>
-    <ProjectReference Include="..\RabbitExpress.ExampleShared\RabbitExpress.ExampleShared.csproj" />
-  </ItemGroup>
-	
-  <ItemGroup>
-    <None Update="appsettings.json">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </None>
-  </ItemGroup>
-
-</Project>
