--- RabbitExpress.JsonSerializer/RabbitExpress.JsonSerializer.csproj
@@ -1,27 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-    <Authors>Rene Windegger</Authors>
-    <Company>Rene Windegger</Company>
-    <Product>RabbitExpress</Product>
-    <Description>A simple JSON serializer implementation for usage in RabbitExpress.</Description>
-    <PackageLicenseUrl>https://github.com/rwindegger/RabbitExpress/blob/master/LICENSE</PackageLicenseUrl>
-    <PackageProjectUrl>https://rwindegger.github.io/RabbitExpress/</PackageProjectUrl>
-    <RepositoryUrl>https://github.com/rwindegger/RabbitExpress</RepositoryUrl>
-    <PackageTags>rabbitmq rabbitmq-client</PackageTags>
-	<Version>99.99.99.99</Version>
-	<PackageVersion>99.99.99.99</PackageVersion>
-	<AssemblyVersion>99.99.99.99</AssemblyVersion>
-	<FileVersion>99.99.99.99</FileVersion>
-	<InformationalVersion>99.99.99.99</InformationalVersion>
-	<RootNamespace>RabbitExpress</RootNamespace>
-  </PropertyGroup>
-  <ItemGroup>
-	<PackageReference Include="Newtonsoft.Json" Version="12.0.2" />
-  </ItemGroup>
-  <ItemGroup>
-	<ProjectReference Include="..\RabbitExpress\RabbitExpress.csproj" />
-  </ItemGroup>
-</Project>
