+++ Serializers/RabbitExpress.Serializers.JsonSerializer/JsonSerializer.cs
