+++ Serializers/RabbitExpress.Serializers.JsonSerializer/README.md
@@ -8,13 +8,13 @@ In your csproj add a PackageReference to this package.
 
 ```xml
 <ItemGroup>
-    <PackageReference Include="RabbitExpress.JsonSerializer" Version="*" />
+    <PackageReference Include="RabbitExpress.Serializers.JsonSerializer" Version="1.*" />
 </ItemGroup>
 ```
 
 ## The serializer
 
-This is a very simple serializer to be used in the [RabbitExpress.QueueClient](../RabbitExpress/README.md). The implementation uses [Newtonsoft.Json](https://www.newtonsoft.com/json) to serialize and deserialize the transfered messages.
+This is a very simple serializer to be used in the [RabbitExpress.QueueClient](../../RabbitExpress/README.md). The implementation uses [Newtonsoft.Json](https://www.newtonsoft.com/json) to serialize and deserialize the transfered messages.
 
 ```c-sharp
 public class JsonSerializer : IExpressSerializer
